+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 1.5 (dev)
   * middleware to get better prometheus stat generation (RED/etc.)
   * Golang 1.18.5
+  * allow for metric names in URLs _or_ in body
 
 ## 1.4
   * metric names in message body